Swap dashboard/marketing routes and switch frontend to a   real build
