+++ .env.example
@@ -1,0 +1,19 @@
+# Used by compose.yaml to pass NEXT_PUBLIC_* build args to the frontend
+# image (Docker Compose reads a file named ".env" in this directory
+# automatically for variable substitution).
+#
+# For local development: copy this file to .env (`cp .env.example .env`)
+# and `docker compose up` will pick these values up automatically.
+#
+# For staging/production: do NOT create a .env from this file. Instead,
+# export NEXT_PUBLIC_BASE_URL and NEXT_PUBLIC_GOOGLE_CLIENT_ID as real
+# environment variables in whatever runs `docker compose up` (CI/CD
+# config, systemd unit, platform env settings, etc.), pointing at the
+# real backend URL and the production Google OAuth client ID. Shell-
+# exported variables take precedence over a .env file, but skipping the
+# .env file entirely means a forgotten override fails loudly (Compose
+# warns about an unset variable) instead of silently falling back to
+# these local-dev values.
+
+NEXT_PUBLIC_BASE_URL=http://localhost:9000
+NEXT_PUBLIC_GOOGLE_CLIENT_ID=""
