+++ web/src/app/dashboard/page.tsx
@@ -1,0 +1,22 @@
+import PageLayout from "@/components/layouts/PageLayout";
+import TopBoard from "@/components/home/TopBoard";
+import LinearChart from "@/components/home/LinearChart";
+import LatestExpenses from "@/components/home/LatestExpenses";
+import ImportCard from "@/components/home/ImportCard";
+import styles from "@/components/CustomStyles";
+import { ProtectedRoute } from "@/components/helpers/RouteProtection";
+
+export default function Home() {
+  return (
+    <ProtectedRoute>
+      <PageLayout>
+        <main className={styles.pagePad}>
+          <TopBoard data-testid="top-board" />
+          <ImportCard />
+          <LinearChart />
+          <LatestExpenses />
+        </main>
+      </PageLayout>
+    </ProtectedRoute>
+  );
+}
