+++ web/src/components/signin/Form.tsx
@@ -39,7 +39,7 @@ const Form = () => {
           localStorage.setItem("Expendit-userID", JSON.stringify(user_id));
           localStorage.setItem("Expendit-user", JSON.stringify(data.user));
           localStorage.setItem("ExpenditLoggedIn", JSON.stringify(true));
-          router.push("/");
+          router.push("/dashboard");
         }
       })
       .catch(err => {
